+++ R/rxSolve.R
@@ -33,7 +33,6 @@ rxSolve.monolix2rx <- function(object, params = NULL, events = NULL,
             }
             else if (!is.null(object$dfObs)) {
                 dfObs <- object$dfObs
-                dfObs <- object$meta$dfObs
                 .minfo(paste0("using dfObs=", dfObs, " from Monolix"))
             }
         }
@@ -43,7 +42,7 @@ rxSolve.monolix2rx <- function(object, params = NULL, events = NULL,
                 .minfo(paste0("using thetaMat from Monolix"))
             }
             else if (!is.null(object$thetaMat)) {
-                thetaMat <- object$meta$thetaMat
+                thetaMat <- object$thetaMat
                 .minfo(paste0("using thetaMat from Monolix"))
             }
         }